chore: move verification flow test seperate module
